Move grades report to new djangoapp.
