+++ lms/envs/common.py
@@ -2029,6 +2029,7 @@ INSTALLED_APPS = (
     'bulk_email',
     'branding',
     'lms.djangoapps.grades.apps.GradesConfig',
+    'lms.djangoapps.grades_report',
 
     # Student support tools
     'support',
